refactor(web): split view and review routes from ip comments
