+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/interested-party-comments.controller.js
@@ -1,10 +1,5 @@
 import logger from '#lib/logger.js';
-import { addNotificationBannerToSession } from '#lib/session-utilities.js';
-import {
-	interestedPartyCommentsPage,
-	reviewInterestedPartyCommentPage,
-	viewInterestedPartyCommentPage
-} from './interested-party-comments.mapper.js';
+import { interestedPartyCommentsPage } from './interested-party-comments.mapper.js';
 import * as interestedPartyCommentsService from './interested-party-comments.service.js';
 
 /**
@@ -61,96 +56,6 @@ export const renderInterestedPartyComments = async (request, response) => {
 			errors
 		});
 	} catch (error) {
-		logger.error(error);
-		return response.status(500).render('app/500.njk');
-	}
-};
-
-/**
- *
- * @param {import('@pins/express/types/express.js').Request} request
- * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
- */
-export async function renderViewInterestedPartyComment(request, response) {
-	const { errors, currentComment } = request;
-
-	if (!currentComment) {
-		return response.status(404).render('app/404.njk');
-	}
-
-	const pageContent = viewInterestedPartyCommentPage(request.currentAppeal, request.currentComment);
-
-	return response.status(200).render('patterns/display-page.pattern.njk', {
-		errors,
-		pageContent
-	});
-}
-
-/**
- * @param {import('@pins/express/types/express.js').Request} request
- * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
- */
-export async function renderReviewInterestedPartyComment(request, response) {
-	const { errors, currentComment } = request;
-
-	if (!currentComment) {
-		return response.status(404).render('app/404.njk');
-	}
-
-	const pageContent = reviewInterestedPartyCommentPage(request.currentAppeal, currentComment);
-
-	return response.status(200).render('patterns/change-page.pattern.njk', {
-		errors,
-		pageContent
-	});
-}
-
-/**
- * @param {import('@pins/express/types/express.js').Request} request
- * @param {import('@pins/express/types/express.js').RenderedResponse<any, any, Number>} response
- */
-export const renderPostReviewInterestedPartyComment = async (request, response) => {
-	try {
-		const {
-			errors,
-			currentAppeal,
-			params: { appealId, commentId },
-			body: { status },
-			apiClient,
-			session
-		} = request;
-
-		if (!currentAppeal) {
-			logger.error('Current appeal not found.');
-			return response.status(500).render('app/500.njk');
-		}
-
-		if (errors) {
-			const pageContent = reviewInterestedPartyCommentPage(
-				request.currentAppeal,
-				request.currentComment
-			);
-
-			return response.status(200).render('patterns/change-page.pattern.njk', {
-				errors,
-				pageContent
-			});
-		}
-
-		await interestedPartyCommentsService.patchInterestedPartyCommentStatus(
-			apiClient,
-			appealId,
-			commentId,
-			status
-		);
-
-		addNotificationBannerToSession(session, 'interestedPartyCommentsValidSuccess', appealId);
-
-		return response.redirect(
-			`/appeals-service/appeal-details/${appealId}/interested-party-comments`
-		);
-	} catch (error) {
-		console.log('🚀 ~ renderPostReviewInterestedPartyComment ~ error:', error);
 		logger.error(error);
 		return response.status(500).render('app/500.njk');
 	}
