+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/interested-party-comments.router.js
@@ -1,29 +1,14 @@
-import { Router as createRouter } from 'express';
 import { asyncHandler } from '@pins/express';
+import { Router as createRouter } from 'express';
 import { validateAppeal } from '../appeal-details.middleware.js';
-import { validateComment } from './interested-party-comments.middleware.js';
 import addIpCommentRouter from './add-ip-comment/add-ip-comment.router.js';
-import * as validators from './interested-party-comments.validators.js';
+import viewAndReviewIpCommentRouter from './view-and-review/view-and-review.router.js';
 import * as controller from './interested-party-comments.controller.js';
 
 const router = createRouter({ mergeParams: true });
 
-router.use('/add', validateAppeal, addIpCommentRouter);
-
+router.use('/', viewAndReviewIpCommentRouter);
 router.route('/').get(validateAppeal, asyncHandler(controller.renderInterestedPartyComments));
-
-router
-	.route('/:commentId/view')
-	.get(validateAppeal, validateComment, asyncHandler(controller.renderViewInterestedPartyComment));
-
-router
-	.route('/:commentId/review')
-	.get(validateAppeal, validateComment, asyncHandler(controller.renderReviewInterestedPartyComment))
-	.post(
-		validateAppeal,
-		validateComment,
-		validators.validateReviewComment,
-		asyncHandler(controller.renderPostReviewInterestedPartyComment)
-	);
+router.use('/add', validateAppeal, addIpCommentRouter);
 
 export default router;
