+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/interested-party-comments.service.js
@@ -28,26 +28,3 @@ export const getInterestedPartyComments = (
 
 	return apiClient.get(url).json();
 };
-
-/**
- * @param {import('got').Got} apiClient
- * @param {string} appealId
- * @param {string} commentId
- * */
-export const getInterestedPartyComment = (apiClient, appealId, commentId) =>
-	apiClient.get(`appeals/${appealId}/reps/${commentId}`).json();
-
-/**
- * @param {import('got').Got} apiClient
- * @param {string} appealId
- * @param {string} commentId
- * @param {string} status
- * */
-export const patchInterestedPartyCommentStatus = (apiClient, appealId, commentId, status) =>
-	apiClient
-		.patch(`appeals/${appealId}/reps/${commentId}/status`, {
-			json: {
-				status
-			}
-		})
-		.json();
